+++ HotAppsLauncher/main.swift
@@ -13,36 +13,21 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     func applicationDidFinishLaunching(_ aNotification: Notification) {
         let mainBundleID = "com.skyrilhd.HotApps"
 
-        // Check if HotApps is running
-        let isRunning = !NSWorkspace.shared.runningApplications.filter { $0.bundleIdentifier == mainBundleID }.isEmpty
+        let runningApps = NSWorkspace.shared.runningApplications
+        let isRunning = runningApps.contains {
+            $0.bundleIdentifier == mainBundleID
+        }
 
         if !isRunning {
-            DistributedNotificationCenter.default().addObserver(self, selector: #selector(self.terminate),
-                                                                name: Notification.Name("killLauncher"),
-                                                                object: mainBundleID)
-
-            // Get HotApps path
-            let launcherPath = Bundle.main.bundlePath as NSString
-            var launcherPathArr = launcherPath.pathComponents
-            launcherPathArr.removeLast()
-            launcherPathArr.removeLast()
-            launcherPathArr.removeLast()
-            launcherPathArr.removeLast()
-            let mainPath = NSString.path(withComponents: launcherPathArr)
-            let mainURL = URL(fileURLWithPath: mainPath)
-
-            if #available(macOS 10.15, *) {
-                NSWorkspace.shared.openApplication(at: mainURL, configuration: NSWorkspace.OpenConfiguration())
-            } else {
-                // Fallback on earlier versions
-                NSWorkspace.shared.open(mainURL)
+            var path = Bundle.main.bundlePath as NSString
+            for _ in 1...4 {
+                path = path.deletingLastPathComponent as NSString
             }
-        } else {
-            self.terminate()
-        }
-    }
 
-    @objc func terminate() {
+            let mainPathString = path as String
+            NSWorkspace.shared.launchApplication(mainPathString)
+        }
+
         NSApp.terminate(nil)
     }
 }
@@ -51,4 +36,4 @@ let app = NSApplication.shared
 let delegate = AppDelegate()
 app.delegate = delegate
 
-NSApplication.shared.run()
+_ = NSApplicationMain(CommandLine.argc, CommandLine.unsafeArgv)