+++ CHANGELOG.md
@@ -4,6 +4,7 @@ HotApps Changelog
 - improved corner detection
 - fixed a bug where a HotApp would hide/unhide even though the pointer was not at corner anymore
 - more performance improvements
+- fixed 'Launch at Login'
 
 
 #### 0.4